Split Menu Drawer Component From Chat Room Component
