+++ packages/client/src/pages/chatroom/MenuDrawer/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MenuDrawer'
